+++ Data/Types/ProductMap.cs
@@ -16,7 +16,7 @@ namespace Data.Types
 
             builder.Property(i => i.Price)
             .HasColumnName("price")
-            .HasColumnType("DOUBLE")
+            .HasColumnType("DECIMAL")
             .IsRequired();
             
             builder.Property(i => i.Name)
@@ -32,7 +32,7 @@ namespace Data.Types
 
             builder.Property(i => i.QuantityKgSold)
             .HasColumnName("quantityKgSold")
-            .HasColumnType("DOUBLE")
+            .HasColumnType("DECIMAL")
             .IsRequired();
 
             builder.HasOne(x => x.Item)
